Django_User_Authentication/user-profile-with-django-master/requirements.txt
Django==2.1.7
django-appconf==1.0.3
django-imagekit==4.0.2
pilkit==2.0
Pillow==5.4.1
pytz==2018.9
six==1.12.0
